Split Moctree.js
